Graphiql bool as env var and moved graphql to TLD
